Rename example files.
