+++ examples/gaussian_diagcov_legacy.py
